Update player count types in Coinflip program to u8 for better memory efficiency
Refactor the max_players and min_players fields in OracleConfig, GameConfig, and related instructions to use u8 instead of u16. This change optimizes memory usage and aligns with the realistic player limits in the game.

+++ programs/coinflip/src/instructions/initialize_game.rs
@@ -6,8 +6,8 @@ pub fn handler(
     ctx: Context<super::InitializeGame>,
     game_type: GameType,
     amount: u64,
-    max_players: u16,
-    min_players: u16,
+    max_players: u8,
+    min_players: u8,
     timeout: u32,
     is_private: bool,
 ) -> Result<()> {
